+++ great_expectations_cloud/agent/actions/__init__.py
@@ -19,4 +19,4 @@ from great_expectations_cloud.agent.actions.run_checkpoint import RunCheckpointA
 from great_expectations_cloud.agent.actions.run_column_descriptive_metrics_action import (
     ColumnDescriptiveMetricsAction,
 )
-from great_expectations_cloud.agent.actions.run_metrics_list_action import MetricsListAction
+from great_expectations_cloud.agent.actions.run_metric_list_action import MetricsListAction
